fixed size_t warnings.
git-svn-id: svn://136.177.114.72/svn_GW/phastpp/trunk@4741 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/phastinput/Shapefiles/dbfopen.c
@@ -330,7 +330,7 @@ DBFOpen(const char *pszFilename, const char *pszAccess)
 /* -------------------------------------------------------------------- */
 	pszBasename = (char *) malloc(strlen(pszFilename) + 5);
 	strcpy(pszBasename, pszFilename);
-	for (i = strlen(pszBasename) - 1;
+	for (i = (int) strlen(pszBasename) - 1;
 		 i > 0 && pszBasename[i] != '.' && pszBasename[i] != '/'
 		 && pszBasename[i] != '\\'; i--)
 	{
@@ -522,7 +522,7 @@ DBFCreate(const char *pszFilename)
 /* -------------------------------------------------------------------- */
 	pszBasename = (char *) malloc(strlen(pszFilename) + 5);
 	strcpy(pszBasename, pszFilename);
-	for (i = strlen(pszBasename) - 1;
+	for (i = (int) strlen(pszBasename) - 1;
 		 i > 0 && pszBasename[i] != '.' && pszBasename[i] != '/'
 		 && pszBasename[i] != '\\'; i--)
 	{
@@ -1112,7 +1112,7 @@ DBFWriteAttribute(DBFHandle psDBF, int hEntity, int iField, void *pValue)
 		{
 			memset(pabyRec + psDBF->panFieldOffset[iField], ' ',
 				   psDBF->panFieldSize[iField]);
-			j = strlen((char *) pValue);
+			j = (int) strlen((char *) pValue);
 		}
 
 		strncpy((char *) (pabyRec + psDBF->panFieldOffset[iField]),
@@ -1188,7 +1188,7 @@ DBFWriteAttributeDirectly(DBFHandle psDBF, int hEntity, int iField,
 	{
 		memset(pabyRec + psDBF->panFieldOffset[iField], ' ',
 			   psDBF->panFieldSize[iField]);
-		j = strlen((char *) pValue);
+		j = (int) strlen((char *) pValue);
 	}
 
 	strncpy((char *) (pabyRec + psDBF->panFieldOffset[iField]),
@@ -1449,7 +1449,7 @@ str_to_upper(char *string)
 	int len;
 	short i = -1;
 
-	len = strlen(string);
+	len = (int) strlen(string);
 
 	while (++i < len)
 		if (isalpha(string[i]) && islower(string[i]))
